// client/src/app/models/entreprise.ts
export interface Entreprise {
    numEntreprise: number | undefined;
    raisonSociale: string;
    nomContact: string;
    nomResp: string;
    rueEntreprise: string;
    cpEntreprise: number | string;
    villeEntreprise: string;
    telEntreprise: string;
    faxEntreprise: string;
    email: string;
    observation: string;
    siteEntreprise: string;
    niveau: string;
    enActivite: number;
}